Update terraform-plugin-sdk to one that supports latest terraform
Tests failed when running with latest terraform (v1.1.0) since they
updated the state file version to a version that the old plugin SDK
didn't accept.

We also enable logging in the terraform cli used by acceptence tests
when running Jenkins, to provide a way to trace which version we used.
The plan is to support testing of specific terraform cli versions in a
future change.